+++ bucket4j/src/main/java/com/linecorp/armeria/server/throttling/bucket4j/package-info.java
@@ -21,6 +21,6 @@
  *
  */
 @NonNullByDefault
-package com.linecorp.armeria.server.throttling.tokenbucket;
+package com.linecorp.armeria.server.throttling.bucket4j;
 
 import com.linecorp.armeria.common.util.NonNullByDefault;
